+++ src/main.tsx
@@ -4,7 +4,7 @@ import { ThemeProvider } from '@emotion/react';
 import { createRoot } from 'react-dom/client';
 
 import GlobalStyles from '@/styles/GlobalStyles';
-import theme from '@/styles/Theme.ts';
+import theme from '@/styles/theme.ts';
 
 import App from './App.tsx';
