+++ src/styles/GlobalStyles.tsx
@@ -1,6 +1,6 @@
 import { css, Global } from '@emotion/react';
 
-import theme from '@/styles/Theme';
+import theme from '@/styles/theme';
 
 const GlobalStyles = () => {
   return <Global styles={baseStyles} />;
