+++ src/types/emotion.d.ts
@@ -1,5 +1,5 @@
 import '@emotion/react';
-import { ColorsType, FontSizeType } from '@/styles/Theme';
+import { ColorsType, FontSizeType } from '@/styles/theme';
 
 declare module '@emotion/react' {
   export interface Theme {
